requirements.txt
fastapi>=0.104.1
gunicorn>=21.2.0
uvicorn>=0.24.0
jinja2>=3.1.2
pyyaml>=6.0.1
paramiko>=3.3.1